Add the channelPersonal setting.
If the channel settings is true, and channelPersonal is also
true, the bot will respond to the user who pinged in the channel,
instead of to the channel itself (by replacing "$(channel)" in
the response with userName instead of channelName).
The reminder setting will also be applied to the individual who
pinged, instead of to the channel.

This is a very useful setting in organizations with large
channels where people join for a short while to ping someone, and
might not read the last few hours of scrollback to see that
the bot replied to someone else earlier.