Add the mirrorPings setting.
With this setting turned on, all the pings (and the
associated messages), will be copied to your personal slack
channel (the one called "@username (you)" in Slack).

With mirror pings you get all pings saved in a central location.
It's a good way to browse through your pings while on vacation to
see if any are important enough to actually reply to.
(Ideally, there should be no such pings ever, but ...)